Move comparison functions from datablock to experiment_list
